+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## unreleased
+### Added
+* Added field `outline_buffers` to `vertex.ModelBuffers`.
+* Added type `vertex.OutlineBuffer`.
+
 ## 0.7.0 - 2024-07-22
 ### Added
 * Added type `ChannelAssignmentAttribute` for vertex attribute shader dependencies.